requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
idna==2.6
pytest-runner==2.11.1
PyYAML==3.12
requests==2.18.4
urllib3==1.22
usgs-datatools==0.1.0
